requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
Cerberus==1.3.5
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
djangorestframework==3.15.1
greenlet==3.0.3
inflect==7.3.1
isort==5.13.2
kombu==5.4.0
more-itertools==10.3.0
mysqlclient==2.2.4
prompt_toolkit==3.0.47
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
six==1.16.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.52
sqlparse==0.5.0
typeguard==4.3.0
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
